+++ Extractor/Connect/ReverseConnectionSource.cs
@@ -42,7 +42,7 @@ namespace Cognite.OpcUa.Connect
                         oldConnection.Session.Endpoint.Server.ApplicationUri,
                         token
                     );
-                    SessionManager.IncConnects();
+                    ConnectionUtils.Connects.Inc();
 
                     if (reconn == null)
                     {
@@ -55,7 +55,7 @@ namespace Cognite.OpcUa.Connect
                 catch (Exception ex)
                 {
                     log.LogError(ex, "Failed to reconnect to server at {Url}: {Message}", oldConnection.Session.Endpoint.EndpointUrl, ex.Message);
-                    if (SessionManager.ShouldAbandonReconnect(ex) || config.ForceRestart)
+                    if (ConnectionUtils.ShouldAbandonReconnect(ex) || config.ForceRestart)
                     {
                         await sessionManager.CloseSession(oldConnection.Session, token);
                     }
@@ -123,7 +123,7 @@ namespace Cognite.OpcUa.Connect
                     0,
                     identity,
                     null);
-                SessionManager.IncConnects();
+                ConnectionUtils.Connects.Inc();
                 session.DeleteSubscriptionsOnClose = true;
                 return new ConnectResult(new Connection(session, endpointUrl), ConnectType.NewSession);
             }
